requirements.txt
pandas 
numpy 
scikit-learn
matplotlib 
seaborn 
rdkit 
shap
streamlit
python-dotenv